# lib/api-zod/src/generated/types/employeeListItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeListItem {
  empId: number;
  firstName: string;
  lastName: string;
  doodleId: string;
  team: string;
  grade: string;
  earnedLeaves: number;
  sickLeaves: number;
  otherLeaves: number;
  totalLeaves: number;
  elDeficit: number;
  slDeficit: number;
  finalTotal: number;
  /** Working days without EOD submission (0 for exempt teams) */
  lagDays: number;
  hrComment?: string | null;
  pmComment?: string | null;
  isDtn: boolean;
}