# lib/api-zod/src/generated/types/laggingEmployee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface LaggingEmployee {
  empId: number;
  name: string;
  doodleId: string;
  team: string;
  /** Number of working days without EOD submission */
  lagDays: number;
  daysLogged: number;
  expectedDays: number;
}