# lib/api-zod/src/generated/types/laggingResourcesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LaggingResourceCategory } from "./laggingResourceCategory";

export interface LaggingResourcesResponse {
  categories: LaggingResourceCategory[];
}