# lib/api-zod/src/generated/types/leaveAnalyticsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveAnalyticsRow } from "./leaveAnalyticsRow";

export interface LeaveAnalyticsResponse {
  rows: LeaveAnalyticsRow[];
}