# lib/api-zod/src/generated/types/productivityResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductivityResponse {
  /** Overall productivity percentage (0-100) */
  productivityPercent: number;
}