# lib/api-zod/src/generated/types/resourceCountRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourceCountRow {
  category: string;
  ftCount: number;
  dtnCount: number;
  total: number;
}