# lib/api-zod/src/generated/types/trendDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HR Productivity Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendDataPoint {
  /** Day or date label (e.g. Mon, Tue or 2024-01-01) */
  label: string;
  productivityPercent: number;
}